+++ CHANGELOG.md
@@ -1,5 +1,12 @@
+## Version 0.2.2
+* Database path is hadrcoded
+* Added command-line argument to add hosts to database
+* Added command-line argument to list entries from database
+* Added "Shutdown" event processing
+* Fixed pointer error in dba_free()
+
 ## Version 0.2.1
-* SQLite replaced with BerkleyDB
+* SQLite storage replaced with BerkleyDB
 
 ## Version 0.1d
 * Added configurable internal network
@@ -11,7 +18,7 @@
 
 ## Version 0.1b
 * Added command-line arguments parser
-* Fix VMAP index error
+* Fixed VMAP index error
 
 ## Version 0.1a
 * First release